Move --json-output option into passthrough
The only reason I was adding this globally is because of how it
displayed in the help output - but since we're using DDEV for that
directly now we can just add this in the pass through, and not worry
about it.

This means we don't have to respect or remove this in any custom
commands.